==> results_scripts/06 Short SVR train predict.ipynb <==
# SVR

This is Python code to fit SVR models on our time series data, and predict ahead

## 1) First, we are getting our data and environment set up

print('hello and welcome to a functioning script')

print('running')
import time
import math
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os
from sklearn.svm import SVR
from sklearn.model_selection import GridSearchCV, TimeSeriesSplit 
from sklearn.preprocessing import StandardScaler
from sklearn import metrics
%matplotlib inline

os.chdir('/rds/general/user/ag1620/home/summerproj/results_data')
os.getcwd()

dataset_path = "short_1.csv"
df = pd.read_csv(dataset_path,header=0,index_col=0)

df['chunk'] =  pd.to_datetime(df['chunk'], format="%Y-%m-%d %H:%M:%S")
print(df.columns)
print(df.shape)

trainlength = df[df['traintest'].isin(range(0,1))].shape[0]
testlength = df[df['traintest'].isin(range(1,2))].shape[0]

## Choose your variables!
date_time = df['chunk']
df = df[['chunk','texts', 'localtime','weekday', 'tweet_avg', 'retweet_avg', 'likes_avg',
       'temp', 'clouds', 'covid_deaths', 'covid_announcements']]

## TRANSFORM YOUR VARIABLES!
df["texts"] = df["texts"].apply(lambda x: math.sqrt(x))
df["tweet_avg"] = df["tweet_avg"].apply(lambda x: math.log(x+0.1))
df["retweet_avg"] = df["retweet_avg"].apply(lambda x: math.log(x+0.1))
df["likes_avg"] = df["likes_avg"].apply(lambda x: math.log(x+0.1))
print("Updated DataFrame:")
df.head()

# Splitting data into 
    # train
    # test

column_indices = {name: i for i, name in enumerate(df.columns)}

n = len(df)
train = df[:int(trainlength)]
test = df[trainlength:]

num_features = df.shape[1]

## Split Truth and Predictors for train
y_train = train['texts']
print(y_train.shape)
X_train = train[['localtime','weekday',
                 'tweet_avg', 'retweet_avg', 'likes_avg',
                 'temp','clouds',
                 'covid_deaths','covid_announcements'
                ]] #
print(X_train.shape)
y_test = test['texts']
print(y_test.shape)
X_test = test[['localtime','weekday',
                 'tweet_avg', 'retweet_avg', 'likes_avg',
                 'temp','clouds',
                 'covid_deaths','covid_announcements'
              ]]  #
print(X_test.shape)

# Creating dummy variables for categorical variables of local time and weekday
X_train = pd.get_dummies(X_train, columns=['localtime','weekday'])
X_test = pd.get_dummies(X_test, columns=['localtime','weekday'])
X_train

print(X_train.shape)
print(X_test.shape)

# Rescale the dataset
scaler = StandardScaler().fit(X_train)
X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)
# Check data types
print("X_train:" + str(type(X_train)))
print(X_train.shape)
print("y_train:" + str(type(y_train)))
print(y_train.shape)
print("X_test:" + str(type(X_test)))
print(X_test.shape)
print("y_test:" + str(type(y_test)))
print(y_test.shape)
X_train = pd.DataFrame(X_train)
X_test = pd.DataFrame(X_test)

# This code kills performance - all predictors should be aligned
#time = train[['time']].reset_index(drop=True)
#X_train = pd.concat([time,X_train.iloc[:,1:]], axis=1,ignore_index=True)

#This code keeps only the weekday and local
#X_train = X_train.iloc[:,1:]
#X_test = X_test.iloc[:,1:]

X_train

## 3) Fitting Models

## Fit regression model
#svr_rbf = SVR(kernel='rbf', C=100, gamma=0.1, epsilon=.1)
#svr_lin = SVR(kernel='linear', C=100, gamma='auto')
#svr_poly = SVR(kernel='poly', C=100, gamma='auto', degree=3, epsilon=.1, coef0=1)

# Cross Validation 1
tss = TimeSeriesSplit(n_splits=4)
parameters = {'kernel': ['rbf'], # ,'linear','poly'
              'C': [1, 10, 100, 1000, 10000], # range(80,120,20), # Lower = More regularisation
              'gamma': [0.0001, 0.001, 0.01, 0.1], # 'scale','auto' # Lower = samples have larger radius of influence - "smoothness of the forecasting function"
              'epsilon': [0.01, 0.1, 1] } # 
gscv = GridSearchCV(SVR(), parameters, n_jobs=1, cv=tss) # scoring='neg_median_absolute_error' > this makes the predictions far worse

# Default internal scoring strategy is R2
print('running')
#import time
t0 = time.time()

gscv.fit(X=X_train, y=y_train)
print(gscv.best_params_)
svr_model = gscv.best_estimator_

t1 = time.time()
total = t1-t0
print(total)

#import time
t0 = time.time()

y_pred = svr_model.predict(X_test[:12]) ** 2
y_test = y_test[:12]
t1 = time.time()
total = t1-t0
print(total)

## 4) Assessing Performance

# over what time would you like to assess error metrics?
days = 1
t=y_test[:days*12] ** 2
p=y_pred[:days*12]
times=test['chunk'][:days*12]

fig = plt.figure(figsize=(12,6))
plt.plot(times,t,color='b',label='True')
plt.plot(times,p,color='r',label='Preds')
plt.xlabel("Time")
plt.ylabel("Texts")
plt.legend()
plt.show()

# ASSESS ERROR METRICS
import math
# Get error metrics
MAE = metrics.mean_absolute_error(t, p)
MSE = metrics.mean_squared_error(t, p)
RMSE = math.sqrt(MSE)
R2 = metrics.r2_score(t, p)
maxerror = metrics.max_error(t, p)
print(' MAE = {}\n MSE = {}\n RMSE = {}\n R2= {}\n max-error = {}'.format(MAE, MSE, RMSE, R2, maxerror))

# training performance
y_trainpred = svr_model.predict(X_train) ** 2
y_traintruth = y_train ** 2
MAE = metrics.mean_absolute_error(y_traintruth, y_trainpred)
MSE = metrics.mean_squared_error(y_traintruth, y_trainpred)
RMSE = math.sqrt(MSE)
R2 = metrics.r2_score(y_traintruth, y_trainpred)
maxerror = metrics.max_error(y_traintruth, y_trainpred)
print(' MAE = {}\n MSE = {}\n RMSE = {}\n R2= {}\n max-error = {}'.format(MAE, MSE, RMSE, R2, maxerror))

## 5) Save Predictions

# save test set performance dataframe
testchunk = pd.DataFrame(test['chunk']).reset_index(drop=True)
ytest = pd.DataFrame(y_test).reset_index(drop=True) ** 2
ypred = pd.DataFrame(y_pred).reset_index(drop=True)
df_c = pd.concat([testchunk,ytest,ypred], axis=1)
df_c.columns = ['chunk','truth','preds']
df_c.head()
df_c.to_csv('super_short_6_SVR_preds.csv')

# save training set performance dataframe
trainchunk = pd.DataFrame(train['chunk']).reset_index(drop=True)
ytrain = pd.DataFrame(y_train).reset_index(drop=True)  ** 2
ytrainpred = pd.DataFrame(y_trainpred).reset_index(drop=True)
df_c = pd.concat([trainchunk,ytrain,ytrainpred], axis=1)
df_c.columns = ['chunk','truth','preds']
df_c.head()
